Add tracking function to squid.sh, use it to track logs.
This seems unfortunately necessary, as squid does not provide any
documented mechanism for logging to STDOUT. Common practice in other
squid containers seems to be symlinking logs to /dev/stdout.

I've opted to use this method as it respects the logging directives in
the squid configuration, works with the default squid user, and is
generally less "magical" in how it works.